zksync_vm2_interface

Module opcodes

EraVM opcodes.

Structs

Add

Add opcode.

And

And opcode.

AuxHeapRead

AuxHeapRead opcode.

AuxHeapWrite

AuxHeapWrite opcode.

AuxMutating0

AuxMutating0 opcode.

Caller

Caller opcode.

CodeAddress

CodeAddress opcode.

ContextMeta

ContextMeta opcode.

ContextU128

ContextU128 opcode.

Decommit

Decommit opcode.

Delegate

Delegate FarCall mode.

Div

Div opcode.

ErgsLeft

ErgsLeft opcode.

Event

Event opcode.

FarCall

FarCall group of opcodes distinguished by the calling mode (normal, delegate, or mimic).

HeapRead

HeapRead opcode.

HeapWrite

HeapWrite opcode.

IncrementTxNumber

IncrementTxNumber opcode.

Jump

Jump opcode.

L2ToL1Message

L2ToL1Message opcode.

Mimic

Mimic FarCall mode.

Mul

Mul opcode.

NearCall

NearCall opcode.

Nop

Nop opcode.

Normal

Normal Return mode / FarCall mode.

Or

Or opcode.

Panic

Panic Return mode.

PointerAdd

PointerAdd opcode.

PointerPack

PointerPack opcode.

PointerRead

PointerRead opcode.

PointerShrink

PointerShrink opcode.

PointerSub

PointerSub opcode.

PrecompileCall

PrecompileCall opcode.

Ret

Ret group of opcodes distinguished by the return type (normal, panic, or revert).

Revert

Revert Return mode.

RotateLeft

RotateLeft opcode.

RotateRight

RotateRight opcode.

SP

SP opcode.

SetContextU128

SetContextU128 opcode.

ShiftLeft

ShiftLeft opcode.

ShiftRight

ShiftRight opcode.

StaticMemoryRead

StaticMemoryRead opcode.

StaticMemoryWrite

StaticMemoryWrite opcode.

StorageRead

StorageRead opcode.

StorageWrite

StorageWrite opcode.

Sub

Sub opcode.

This

This opcode.

TransientStorageRead

TransientStorageRead opcode.

TransientStorageWrite

TransientStorageWrite opcode.

Xor

Xor opcode.

Traits

TypeLevelCallingMode

Calling mode for the FarCall opcodes.

TypeLevelReturnType

Return type for the Ret opcodes.